changed numsp from int to double, everything works now

+++ bcuser.h
@@ -13,7 +13,7 @@ public:
 	BCUser(int i);
 	~BCUser();
 	int dist; //distance to current source vertex
-	int numsp; //number of shortest paths
+	double numsp; //number of shortest paths
 	Mylist<int> preds; //List of predecessors
 	double delta; //partial BC score
 	double bc; //total BC score
